+++ components/utils/index.js
@@ -1,15 +1,32 @@
 export { BootstrapContainer, wrapBootstrap } from './bootstrap-container.jsx';
+export { mapFromInnerRef, mapToInnerRef } from './forwardref-wrapper.jsx';
+
+const variationCache = {};
 
 export function applyVariations(component, variationMap, props) {
 	let wrappedComponent = component;
 	const filteredProps = { ...props };
 
+	const variations = [];
+
 	for (const variation of Object.keys(variationMap) || []) {
 		if (props.hasOwnProperty(variation)) {
-			wrappedComponent = variationMap[variation](wrappedComponent);
+			variations.push(variation);
 			delete filteredProps[variation];
 		}
 	}
+
+	const cacheKey = JSON.stringify(variations.sort());
+
+	if (variationCache[cacheKey] != null) {
+		return { component: variationCache[cacheKey], filteredProps };
+	}
+
+	for (const variation of variations) {
+		wrappedComponent = variationMap[variation](wrappedComponent);
+	}
+
+	variationCache[cacheKey] = wrappedComponent;
 
 	return { component: wrappedComponent, filteredProps };
 }
@@ -27,6 +44,18 @@ export const debouncedResize = callback => {
 	return {
 		cancel: () => window.removeEventListener('resize', listener),
 	};
+};
+
+export const addSeparator = (items = [], separator = '\u00a0\u00a0•\u00a0\u00a0') => {
+	const result = [];
+	for (const item of items) {
+		if (item) {
+			result.push(item, separator);
+		}
+	}
+
+	result.pop();
+	return result;
 };
 
 export const resetStyles = `